+++ data/src/main/kotlin/nl/codecraft/data/remote/GitHubApiService.kt
@@ -1,0 +1,10 @@
+package nl.codecraft.data.remote
+
+import nl.codecraft.data.remote.dto.RepoDto
+import retrofit2.http.GET
+import retrofit2.http.Path
+
+interface GitHubApiService {
+    @GET("users/{user}/repos")
+    suspend fun getUserRepos(@Path("user") user: String): List<RepoDto>
+}