+++ domain/src/main/kotlin/nl/codecraft/domain/repository/GitHubRepository.kt
@@ -1,0 +1,9 @@
+package nl.codecraft.domain.repository
+
+import kotlinx.coroutines.flow.Flow
+import nl.codecraft.model.Repo
+
+interface GitHubRepository {
+    fun getUserRepos(username: String): Flow<List<Repo>>
+    suspend fun refreshUserRepos(username: String)
+}